Factored out ViewModelRevision
