+++ packages/hub/src/app/users/[username]/models/[slug]/view/page.tsx
@@ -5,7 +5,6 @@ import { useFragment } from "react-relay";
 import { ModelPage$key } from "@/__generated__/ModelPage.graphql";
 import { ModelPageFragment, useModelPageQuery } from "../ModelPage";
 import { ViewModelRevision } from "../ViewModelRevision";
-import { ViewModelRevisionContent } from "../ViewModelRevisionContent";
 
 export default function OuterModelPage({
   params,
@@ -20,12 +19,8 @@ export default function OuterModelPage({
   const model = useFragment<ModelPage$key>(ModelPageFragment, modelRef);
 
   return (
-    <ViewModelRevision
-      revisionRef={model.currentRevision}
-      modelUsername={params.username}
-      modelSlug={params.slug}
-    >
-      <ViewModelRevisionContent revisionRef={model.currentRevision} />
-    </ViewModelRevision>
+    <div className="py-4 px-8">
+      <ViewModelRevision revisionRef={model.currentRevision} />
+    </div>
   );
 }
